refactor: added constants util
