+++ react/lib/util/constants.ts
@@ -1,0 +1,9 @@
+export const FIAT_CURRENCIES = ['USD', 'CAD' ];
+
+export const CRYPTO_CURRENCIES = ['BCH', 'XEC'];
+
+export const DECIMALS: { [key: string]: number } = {
+    BCH: 8,
+    XEC: 2,
+    FIAT: 2,
+};
