+++ react/lib/util/format.ts
@@ -1,11 +1,6 @@
 import currencyFormat from 'currency-formatter';
 import { currency } from './api-client';
-
-export const DECIMALS = {
-  BCH: 8,
-  XEC: 2,
-  FIAT: 2,
-};
+import { DECIMALS } from './constants'
 
 export const amount = (x?: number | null): string | undefined => {
   const clean = +x!;
